requirements.txt
pythonocc-core
numpy
ezdxf
fastapi
uvicorn
pydantic
numpy-stl
pillow
matplotlib
pyvista
vtk
